refactor(UI-1269): simplify DashboardProjectsTableRow
