+++ src/interfaces/components/dashboard.interface.ts
@@ -1,0 +1,23 @@
+import { NavigateFunction } from "react-router-dom";
+
+import { DeploymentStateVariant } from "@src/enums";
+import { DashboardProjectWithStats } from "@src/types/models";
+
+export interface WelcomeVideoCardProps {
+	description?: string;
+	title: React.ReactNode;
+	onPlay?: () => void;
+	image: string;
+}
+
+export interface DashboardProjectsTableRowProps extends DashboardProjectWithStats {
+	displayDeleteModal: (
+		status: DeploymentStateVariant,
+		deploymentId: string,
+		projectId: string,
+		projectName: string
+	) => void;
+	downloadProjectExport: (projectId: string) => void;
+	handelDeactivateDeployment: (deploymentId: string) => Promise<void>;
+	navigate: NavigateFunction;
+}
