+++ src/interfaces/components/index.ts
@@ -58,4 +58,7 @@ export type { Toast, ToasterTypes } from "@interfaces/components/toast.interface
 export type { ToggleProps } from "@interfaces/components/toggle.interface";
 export type { TypographyProps } from "@interfaces/components/typography.interface";
 export type { UserFeedbackFormProps } from "@interfaces/components/userFeedback.interface";
-export type { WelcomeVideoCardProps } from "@interfaces/components/welcomePageCard.interface";
+export type {
+	WelcomeVideoCardProps,
+	DashboardProjectsTableRowProps,
+} from "@src/interfaces/components/dashboard.interface";
