--- src/interfaces/components/welcomePageCard.interface.ts
@@ -1,6 +1,0 @@
-export interface WelcomeVideoCardProps {
-	description?: string;
-	title: React.ReactNode;
-	onPlay?: () => void;
-	image: string;
-}
